usb: rmnet: open BAM data channel only when ready
Add platform driver to be notified of when the BAM channel is ready to
be opened. This fixes the case of when a USB cable is connected on
device startup and the RmNet driver fails when attempting to open the
BAM MUX channel prematurely due to the configuration delay of around 40
seconds. Thus until the channel is ready, do not allow the call to open.
Also fix port cleanup for both data and SMD control channels.

Change-Id: Idba84504c41596b8fa9f55e731a59b41d7906966
Signed-off-by: Jack Pham <jackp@codeaurora.org>
Signed-off-by: Vamsi Krishna <vskrishn@codeaurora.org>