cpufreq : Fix crash input event handler on governor switch
When an input event and CPU1 hotplug occurs at the same
time, a race condition can happen between cpufreq and
cpu hotplug resulting in a kernel panic.
When a hotplug operation is underway on CPU1,
__cpufreq_remove_dev is called. __cpufreq_remove_dev gets
the lock, proceeds to free the policy info of the current
governor and calls the ondemand governor with a STOP event.
If an input event occurs around the same time, it waits
for __cpufreq_remove_dev to release the lock. Meanwhile,
the ondemand governor clears some of its settings, but
fails to clear its local current policy. As soon as the
input event handler acquires the lock, it now uses the
invalid ondemand current policy handle to make frequency
changes. This results in a kernel panic.

Skip making frequency changes if this condition happens.

CRs-fixed: 327622,319348
Signed-off-by: Anitha Anand <anithas@codeaurora.org>
(cherry picked from commit 3dd6509401d44d16231002fbe8674f4e5f29a1a2)

Change-Id: I37a3d97bc74c828669e71af52d8f4012f108e917
Signed-off-by: Swetha Basineni <sbasin@codeaurora.org>